Refactor our prototype veteran objects to common
Allows referencing of the object from server tests as well as client
code.
