+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "nodemon npm start",
     "karma": "karma start",
-    "test:server": "mocha --recursive ./test/server",
+    "test:server": "mocha --recursive ./test/server ./test/common",
     "lint": "eslint --ext .js --ext .jsx lib src test",
     "start": "node src/server.js",
     "test": "npm run lint && npm run karma -- --single-run",
