+++ src/common/fields.js
@@ -1,4 +1,4 @@
-import _ from 'lodash';
+const _ = require('lodash');
 
 /**
  * Represents an input field value.
@@ -7,14 +7,15 @@ import _ from 'lodash';
  * Without this state, it is extremely hard (impossible?) to write UI with required fields where
  * the initial empty state does not get marked as a distracting error.
  */
-export function makeField(value, dirty = false) {
+function makeField(value, optionalDirty) {
+  const dirty = optionalDirty === undefined ? false : optionalDirty;
   return { value, dirty };
 }
 
 /**
  * Walks through an object hierarchy of fields and marks everything dirty.
  */
-export function dirtyAllFields(field) {
+function dirtyAllFields(field) {
   if (_.keys(field).length === 2 && _.has(field, 'value') && _.has(field, 'dirty')) {
     return makeField(field.value, true);
   } else if (_.isPlainObject(field)) {
@@ -25,3 +26,5 @@ export function dirtyAllFields(field) {
 
   return field;
 }
+
+module.exports = { makeField, dirtyAllFields };
