+++ test/client/components/questions/Phone.spec.jsx
@@ -4,7 +4,7 @@ import sinon from 'sinon';
 import { expect } from 'chai';
 
 import Phone from '../../../../src/client/components/questions/Phone';
-import { makeField } from '../../../../src/client/reducers/fields';
+import { makeField } from '../../../../src/common/fields';
 
 describe('<Phone>', () => {
   describe('propTypes', () => {
